Included the JS function to replace image SVGs to inline onLoad.
Now it'll let us theme our SVG using CSS as if it is inline and we can save the SVGs as files and use it without polluting the template.

This is best of both worlds for icons. If you're using a lot of SVG on a page this might include some overhead. That's the reason I don't want to write it as a global script that converts all SVG.

This script can be used sparingly on places only when it is required.