Finished style changes for the same radio button version of the same markup
With little CSS effort now we've achieved both the radio button version and the toggle button styles with same markup structure.
This reduces conditional code (branching) and that'll always make the code easier to read and understand. Just one class name and you get both toggle and checkbox version of the questions