Implemented a template based DOM loading approach to reduce direct DOM manipulations needed.
This commit includes changes to a hacky state based routing implementation and a template based UI loading mechanism.
These are far from even a acceptable solution, but can be a good starting point.

Need to remove the number based routing with an actual router using history API.

Need some changes in the markup and styles to work better for the template based use case.