Finished forward and backward navigations without persisting data
This is still using the magic number based router, it is working good but can easily go out of sync when there is an error in processing the next/back actions.
Next step is to implement entering and persisting data and then changing this router with a History API based router.