Modified the JSON file to create a unified format for the questions and answers
Spent almost an hour debugging why the included template is not passing right value but never thought of having a closer look at the JSON file.
I assumed the JSON file had the same format for the same type of question, rookie mistake.

Now standardised the data format and it makes it easier to handling templating logic