Feat: Finally figured out how to write unit tests without having to refactor with modules
Using Jasmine's spec runner, removed the dependencies and loading jasmine from a CDN. This let us run tests using a browser.

I am not sure if this can be used in a build pipeline, but for that purpose I might use a test runner such as Karma. I haven't configured Karma on my own and used it before.
I don't want to spend time on configuring a test runner and optimising for the build pipeline for a project like this. So I used specRunner approach for testing the code.