Installed Karma and Jasmine and ported the spec runner test to module based test
This was harder than I expected with conflicts from node, jasmine and karma due to my limited experience of configuring them on my own.
I avoided including a build step that can enable lot of options such as babel for compiling the code to ES5 to run tests normally without modules.

But configuring a build process is a overkill especially since the actual app is done already.
I spent 4 hours on nooks and corners of these docs to find out how to properly load the tests and the specs as modules and load the assets.

The effort was worth it because tests are running without having to make any changes to the app's implementation, often times we have to do things to make our testing framework happy.
I collected few blog ideas out of this effort and that is more than enough justification for spending that much time.